11w

+++ services/interal/model/Answer.go
@@ -1,12 +1,12 @@
 package model
 
 type Answer struct {
-	Id int
-	Task_id int
+	Id int64
+	Task_id int64
 	Answer int
 }
 
-func NewAnswer(task_id int, answer int) *Answer {
+func NewAnswer(task_id int64, answer int) *Answer {
 	return &Answer{
 		Task_id: task_id,
 		Answer: answer,
